Move events file to top level of models
