+++ src/isar/state_machine/states/returning_home.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING, Callable, List, Optional
 
 from isar.eventhandlers.eventhandler import EventHandlerBase, EventHandlerMapping
 from isar.models.communication.message import StartMissionMessage
-from isar.models.communication.queues.events import Event
+from isar.models.events import Event
 from isar.state_machine.utils.common_event_handlers import (
     mission_failed_event_handler,
     mission_started_event_handler,
