+++ src/isar/state_machine/states/unknown_status.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING, Callable, List, Optional
 
 from isar.eventhandlers.eventhandler import EventHandlerBase, EventHandlerMapping
-from isar.models.communication.queues.events import Event
+from isar.models.events import Event
 from isar.state_machine.utils.common_event_handlers import stop_mission_event_handler
 from robot_interface.models.mission.status import RobotStatus
