+++ tests/isar/apis/scheduler/test_scheduler_router.py
@@ -12,7 +12,7 @@ from isar.apis.models.models import ControlMissionResponse
 from isar.apis.models.start_mission_definition import StopMissionDefinition
 from isar.mission_planner.local_planner import LocalPlanner
 from isar.mission_planner.mission_planner_interface import MissionPlannerError
-from isar.models.communication.queues.events import EventTimeoutError
+from isar.models.events import EventTimeoutError
 from isar.services.utilities.scheduling_utilities import SchedulingUtilities
 from isar.state_machine.states_enum import States
 from tests.test_double.mission_definition import DummyMissionDefinition
