+++ tests/isar/services/utilities/test_scheduling_utilities.py
@@ -4,7 +4,7 @@ import pytest
 from fastapi import HTTPException
 from pytest_mock import MockerFixture
 
-from isar.models.communication.queues.events import APIEvent, Event, EventTimeoutError
+from isar.models.events import APIEvent, Event, EventTimeoutError
 from isar.services.utilities.scheduling_utilities import SchedulingUtilities
 from isar.state_machine.states_enum import States
 from tests.test_double.mission_definition import DummyMissionDefinition
